+++ src/idtrackerai/postprocess/trajectories_creation.py
@@ -32,9 +32,7 @@ def trajectories_API(
     video.create_trajectories_timer.start()
     create_dir(video.trajectories_folder, remove_existing=True)
 
-    trajectories = produce_output_dict(
-        list_of_blobs.blobs_in_video, video, conf.save_areas
-    )
+    trajectories = produce_output_dict(list_of_blobs.blobs_in_video, video)
 
     trajectories_file = video.trajectories_folder / (
         "trajectories_wo_identification.npy"
@@ -106,7 +104,7 @@ def interpolate_crossings(video: Video, list_of_fragments: ListOfFragments):
         f"{trajectories_wo_gaps_file}"
     )
     trajectories_wo_gaps = produce_output_dict(
-        list_of_blobs_no_gaps.blobs_in_video, video, conf.save_areas
+        list_of_blobs_no_gaps.blobs_in_video, video
     )
 
     np.save(trajectories_wo_gaps_file, trajectories_wo_gaps)  # type: ignore
@@ -132,9 +130,7 @@ def interpolate_crossings(video: Video, list_of_fragments: ListOfFragments):
         list_of_blobs, list_of_blobs_no_gaps
     )
     trajectories_file = video.trajectories_folder / "trajectories.npy"
-    trajectories = produce_output_dict(
-        list_of_blobs.blobs_in_video, video, conf.save_areas
-    )
+    trajectories = produce_output_dict(list_of_blobs.blobs_in_video, video)
     np.save(trajectories_file, trajectories)  # type: ignore
     if conf.CONVERT_TRAJECTORIES_TO_CSV_AND_JSON:
         convert_trajectories_file_to_csv_and_json(trajectories_file)